refactor!: split bus primitives into shared eigsep_redis package
Extract generic Redis primitives (connection, streams, metadata,
heartbeat, status, config) from eigsep_observing.EigsepRedis into a
new in-repo eigsep_redis package. Observing-specific corr/VNA data
methods remain on EigsepObsRedis(EigsepRedis). Dummies split the same
way: DummyEigsepRedis in eigsep_redis.testing for bus-only tests,
DummyEigsepObsRedis in eigsep_observing.testing for observing tests.

This makes the upcoming monorepo migration mechanical: picohost will
depend only on eigsep_redis, and the new package can be git-mv'd out
as-is with only a pyproject.toml to drop beside it.

BREAKING CHANGE: callers using corr/VNA methods must switch from
EigsepRedis to EigsepObsRedis. eigsep_observing.EigsepRedis now
resolves to the base bus class (no corr/VNA methods).
