+++ scripts/panda_observe.py
@@ -2,7 +2,7 @@ from argparse import ArgumentParser
 import logging
 from threading import Thread
 
-from eigsep_observing import EigsepRedis, PandaClient
+from eigsep_observing import EigsepObsRedis, PandaClient
 from eigsep_observing.testing import DummyPandaClient
 from eigsep_observing.utils import configure_eig_logger
 
@@ -19,11 +19,11 @@ args = parser.parse_args()
 
 if args.dummy:
     logger.warning("Running in DUMMY mode, no hardware will be used.")
-    redis = EigsepRedis(host="localhost", port=6380)
+    redis = EigsepObsRedis(host="localhost", port=6380)
     redis.reset()  # reset test redis database
     client = DummyPandaClient(redis)
 else:
-    redis = EigsepRedis(host="localhost", port=6379)
+    redis = EigsepObsRedis(host="localhost", port=6379)
     client = PandaClient(redis)
 
 logger.info(f"Client configuration: {client.cfg}")
