+++ src/eigsep_observing/__init__.py
@@ -1,7 +1,9 @@
 __author__ = "Christian Hellum Bye"
 __version__ = "1.0.0"
 
-from .eig_redis import EigsepRedis
+from eigsep_redis import EigsepRedis
+
+from .eig_redis import EigsepObsRedis
 from .client import PandaClient
 from .observer import EigObserver
 from .fpga import EigsepFpga
