+++ src/eigsep_observing/testing/__init__.py
@@ -1,5 +1,16 @@
+from eigsep_redis.testing import DummyEigsepRedis
+
 from . import utils
 from .client import DummyPandaClient
-from .eig_redis import DummyEigsepRedis
+from .eig_redis import DummyEigsepObsRedis
 from .fpga import DummyEigsepFpga
 from .observer import DummyEigObserver
+
+__all__ = [
+    "DummyEigObserver",
+    "DummyEigsepFpga",
+    "DummyEigsepObsRedis",
+    "DummyEigsepRedis",
+    "DummyPandaClient",
+    "utils",
+]
