+++ src/eigsep_redis/__init__.py
@@ -1,0 +1,3 @@
+from .eig_redis import EigsepRedis
+
+__all__ = ["EigsepRedis"]
