+++ src/eigsep_redis/testing/__init__.py
@@ -1,0 +1,3 @@
+from .eig_redis import DummyEigsepRedis
+
+__all__ = ["DummyEigsepRedis"]
